// RMG/ClientApp/src/app/models/ReportsEPMAttribute.ts
export class ReportsEPMAttribute {
    project_Assign_ID(arg0: string, project_Assign_ID: any): any {
        throw new Error("Method not implemented.");
    }
  project_assign_id: string;

  emp_Id: string;
  emp_Name: string;
  project_Name: string;
  project_Code: string;
  coe: string;
  edgePractice: string;
  project_Manager_Employee_Id: string;
  project_Manager_Emp_Name: string;
  assign_Project_StartDate: string;
  assign_Project_EndDate: string;
  billable: string;
  billing_Percentage: string;
  location: string;
  onsite: string;


}